+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPage.epics.js
@@ -207,3 +207,4 @@ export const openEnrollmentPageEpic = (action$: InputObservable, store: ReduxSto
         },
         ),
     );
+