+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/DataEntry/TrackedEntityInstance/DataEntryTrackedEntityInstance.component.js
@@ -6,38 +6,41 @@ import { DATA_ENTRY_ID } from '../../registerTei.const';
 import teiClasses from './trackedEntityInstance.module.css';
 import { TeiRegistrationEntry } from '../../../../../DataEntries';
 import type { Props } from './dataEntryTrackedEntityInstance.types';
+import { useCurrentOrgUnitInfo } from '../../../../../../hooks/useCurrentOrgUnitInfo';
 
 const RelationshipTrackedEntityInstancePlain =
-  ({
-      theme,
-      onSave,
-      teiRegistrationMetadata = {},
-      duplicatesReviewPageSize,
-      renderDuplicatesDialogActions,
-      renderDuplicatesCardActions,
-      ExistingUniqueValueDialogActions,
-  }: Props) => {
-      const fieldOptions = { theme, fieldLabelMediaBasedClass: teiClasses.fieldLabelMediaBased };
-      const { trackedEntityType } = teiRegistrationMetadata || {};
-      const trackedEntityTypeNameLC = trackedEntityType.name.toLocaleLowerCase();
+    ({
+        theme,
+        onSave,
+        teiRegistrationMetadata = {},
+        duplicatesReviewPageSize,
+        renderDuplicatesDialogActions,
+        renderDuplicatesCardActions,
+        ExistingUniqueValueDialogActions,
+    }: Props) => {
+        const { id: orgUnitId } = useCurrentOrgUnitInfo();
+        const fieldOptions = { theme, fieldLabelMediaBasedClass: teiClasses.fieldLabelMediaBased };
+        const { trackedEntityType } = teiRegistrationMetadata || {};
+        const trackedEntityTypeNameLC = trackedEntityType.name.toLocaleLowerCase();
 
-      return (
-          // $FlowFixMe - flow error will be resolved when rewriting relationship metadata fetching
-          <TeiRegistrationEntry
-              id={DATA_ENTRY_ID}
-              teiRegistrationMetadata={teiRegistrationMetadata}
-              selectedScopeId={teiRegistrationMetadata.form.id}
-              saveButtonText={i18n.t('Save new {{trackedEntityTypeName}} and link', {
-                  trackedEntityTypeName: trackedEntityTypeNameLC, interpolation: { escapeValue: false },
-              })}
-              fieldOptions={fieldOptions}
-              onSave={onSave}
-              duplicatesReviewPageSize={duplicatesReviewPageSize}
-              renderDuplicatesDialogActions={renderDuplicatesDialogActions}
-              renderDuplicatesCardActions={renderDuplicatesCardActions}
-              ExistingUniqueValueDialogActions={ExistingUniqueValueDialogActions}
-          />
-      );
-  };
+        return (
+            // $FlowFixMe - flow error will be resolved when rewriting relationship metadata fetching
+            <TeiRegistrationEntry
+                id={DATA_ENTRY_ID}
+                orgUnitId={orgUnitId}
+                teiRegistrationMetadata={teiRegistrationMetadata}
+                selectedScopeId={teiRegistrationMetadata.form.id}
+                saveButtonText={i18n.t('Save new {{trackedEntityTypeName}} and link', {
+                    trackedEntityTypeName: trackedEntityTypeNameLC, interpolation: { escapeValue: false },
+                })}
+                fieldOptions={fieldOptions}
+                onSave={onSave}
+                duplicatesReviewPageSize={duplicatesReviewPageSize}
+                renderDuplicatesDialogActions={renderDuplicatesDialogActions}
+                renderDuplicatesCardActions={renderDuplicatesCardActions}
+                ExistingUniqueValueDialogActions={ExistingUniqueValueDialogActions}
+            />
+        );
+    };
 
 export const RelationshipTrackedEntityInstance = withTheme()(RelationshipTrackedEntityInstancePlain);