+++ src/core_modules/capture-core/components/Pages/common/TEIRelationshipsWidget/RegisterTei/RegistrationSection/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { RegistrationSection } from './RegistrationSection.component';
+export { actionTypes } from './registrationSection.actions';