+++ src/core_modules/capture-core/components/Pages/common/TEIRelationshipsWidget/TeiSearch/SearchProgramSelector/SearchProgramSelector.container.js
@@ -1,0 +1,20 @@
+// @flow
+import { connect } from 'react-redux';
+import { SearchProgramSelectorComponent } from './SearchProgramSelector.component';
+import { startSetProgram } from './searchProgramSelector.actions';
+import { getProgramOptions } from './getProgramOptions';
+
+const mapStateToProps = (state: ReduxState, props: Object) => ({
+    programOptions: getProgramOptions(props.selectedTrackedEntityTypeId),
+});
+
+const mapDispatchToProps = (dispatch: ReduxDispatch) => ({
+    onSetProgram: (searchId: string, programId: ?string) => {
+        dispatch(startSetProgram(searchId, programId));
+    },
+});
+
+// $FlowFixMe[missing-annot] automated comment
+export const SearchProgramSelector = connect(mapStateToProps, mapDispatchToProps)(
+    SearchProgramSelectorComponent,
+);