+++ src/core_modules/capture-core/components/WidgetsRelationship/WidgetTrackedEntityRelationship/NewTrackedEntityRelationship/RetrieverModeSelector/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { RetrieverModeSelector } from './RetrieverModeSelector.component';