+++ src/core_modules/capture-core/components/WidgetsRelationship/WidgetTrackedEntityRelationship/NewTrackedEntityRelationship/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { NewTrackedEntityRelationship } from './NewTrackedEntityRelationship.container';